Меняет архитектуру проекта
