+++ js/helpers/show-screen.js
