+++ js/screens/result-lose.js
@@ -1,0 +1,22 @@
+import welcome from './welcome';
+import {getElementFromTemplate} from '../helpers/getElementFromTemplate';
+import {showScreen} from '../helpers/show-screen';
+
+const template = `
+<section class="main main--result">
+  <section class="logo" title="Угадай мелодию"><h1>Угадай мелодию</h1></section>
+
+  <h2 class="title">Вы проиграли</h2>
+  <div class="main-stat">Ничего, вам повезет в следующий раз</div>
+  <span role="button" tabindex="0" class="main-replay">Сыграть ещё раз</span>
+</section>
+`;
+
+const screen = getElementFromTemplate(template);
+const replayButton = screen.querySelector(`.main-replay`);
+
+replayButton.addEventListener(`click`, () => {
+  showScreen(welcome);
+});
+
+export default screen;
